feat(reserve): Change Long to int

+++ src/main/java/com/wanted/preonboarding/core/api/ReserveRequest.java
@@ -3,9 +3,9 @@ package com.wanted.preonboarding.core.api;
 public record ReserveRequest(
         String reserverName,
         String reserverPhone,
-        Long reserverBalance,
-        Long performanceId,
-        Long performanceRound,
+        int reserverBalance,
+        int performanceId,
+        int performanceRound,
         String performanceSeat
 ) {
 }
